Added Code
This version behaves properly regarding how I use memory and seems to be
able to push data fro Temperature, Humidity and Rainfall to nimits.
However, it inexplicably fails after a short time when not connected to
the computer usb so no debug info can be seen to try and find the issue.
Possibly try software serial port but may get RAM issues???